chg: usr: Default share directory in the container was moved to /var/topology to avoid race condition on certains setups where /tmp/ is mounted as tmpfs.
This directory can now be configured using the shared_dir_mount argument in the constructor. In the same way the node exposes this attribute as public
interface. Added to shared_dir_mount other attributes were made public and read only: image, container_id, container_name and shared_dir.

Also, the format of the share_dir in the host changed. Now, the following format will be used:

/tmp/topology/{identifier}_{pid}_{timestamp}