Test TOPOLOGY_ID feature
This feature allows tests to define a TOPOLOGY_ID instead of a
TOPOLOGY in the test module. TOPOLOGY_ID is a string specifying
the basename of a SZN file containing the definition of the
topology. The path where the SZN files are is defined by the value
of --topology-szn-dir pytest option.

Example: TOPOLOGY_ID='myfile', causes topology to look for
the file  {topology-szn-dir}/myfile.szn and obtain the topology
definition from it.